pyproject.toml
[tool.poetry]
name = "trivia-box-bot-tg"
version = "0.1.0"
description = ""
authors = ["Valerian <@VelValerian>"]
readme = "README.md"
packages = [{include = "trivia_box_bot_tg"}]

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.1.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
